requirements.txt
pykrx
pandas
pandas_ta
requests
beautifulsoup4
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
matplotlib
fpdf2
finance-datareader
markdown
schedule
